feat(hum-identity): carve persistent bee identity into a leaf crate
Extract load_or_mint_bee_key / BeeKey / bee_key_path from nest-common
into a standalone hum-identity crate so remote hives can depend on it
without pulling the daemon tree (ensemble/iroh, nest, mcp, etc).

- New crate hum-identity: ids (Hid) + hum-paths + ed25519-dalek + rand.
- hives/common/src/identity.rs becomes a thin re-export shim for
  back-compat; existing nest_common::load_or_mint_bee_key call sites
  keep resolving.
- Registered in workspace members; added to hives/common deps.
