+++ hives/common/Cargo.toml
@@ -7,6 +7,7 @@ description = "Shared building blocks for nests — regex Classifier for the dro
 
 [dependencies]
 hum-paths = { path = "../../hum-paths" }
+hum-identity = { path = "../../hum-identity" }
 drone = { path = "../../drone" }
 ensemble = { path = "../../ensemble" }
 mcp = { path = "../../mcp" }
