+++ hum-identity/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "hum-identity"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+description = "Persistent bee identity: load-or-mint an Ed25519 seed and derive the role-tagged Hid. Leaf crate for remote hives — no daemon, mesh, or nest deps."
+
+[dependencies]
+ids = { path = "../ids" }
+hum-paths = { path = "../hum-paths" }
+ed25519-dalek = { version = "2", features = ["rand_core"] }
+rand = { workspace = true }
+anyhow = { workspace = true }
+tracing = { workspace = true }
+
+[dev-dependencies]
+tempfile = "3"
